refactor: refactor the logic in TransactionView into viewModel
