+++ tokolator/ViewModels/CashierViewModel.swift
@@ -123,4 +123,9 @@ class CashierViewModel: ObservableObject {
             print("Error saving transaction: \(error)")
         }
     }
+
+    func generateDummyData() {
+            let generator = DummyDataGenerator(modelContext: modelContext)
+            generator.generateDummyTransactions()
+        }
 }
